Update retraction speeds and lengths (#283)
* Update retraction speeds and lengths

I changed all default lengths and speeds to both be consistent and have a short default length. Yanking it 10mm back at 60mm/s is just asking for trouble. There's been many clogs because of these defaults. This is a safe startingpoint that might be a little annoying with ooze during next print, but users can tweak this WITHOUT a clogged hotend.

* Update to 5mm@30mm/s